Support lazy initialization of draft columns
This adds `@Column(lazyInitializable:)` for optionalizing draft columns
for lazy initialization.
