+++ Sources/StructuredQueriesMacros/TableMacro.swift
@@ -305,6 +305,9 @@ extension TableMacro: ExtensionMacro {
               }
               isGenerated = true
 
+            case .some(let label) where label.text == "lazyInitializable":
+              break
+
             case let argument?:
               fatalError("Unexpected argument: \(argument)")
             }
@@ -526,6 +529,17 @@ extension TableMacro: ExtensionMacro {
                   node: argument.expression,
                   message: MacroExpansionErrorMessage(
                     "Argument 'generated' is not supported on enum table columns"
+                  )
+                )
+              )
+              continue
+
+            case .some(let label) where label.text == "lazyInitializable":
+              diagnostics.append(
+                Diagnostic(
+                  node: argument.expression,
+                  message: MacroExpansionErrorMessage(
+                    "Argument 'lazyInitializable' is not supported on enum table columns"
                   )
                 )
               )
@@ -790,6 +804,7 @@ extension TableMacro: MemberMacro {
         var isEphemeral = false
         var isExplicitColumn = false
         var isGenerated = false
+        var isLazyInitializable: Bool?
 
         for attribute in property.attributes {
           guard
@@ -857,6 +872,11 @@ extension TableMacro: MemberMacro {
               else { continue }
               isGenerated = true
 
+            case .some(let label) where label.text == "lazyInitializable":
+              isLazyInitializable =
+                argument.expression.as(BooleanLiteralExprSyntax.self)?.literal.tokenKind
+                == .keyword(.true)
+
             case let argument?:
               fatalError("Unexpected argument: \(argument)")
             }
@@ -922,6 +942,15 @@ extension TableMacro: MemberMacro {
         allColumnNames.append(identifier)
         if !isGenerated {
           writableColumns.append(identifier)
+          let lazyInitializableByDefault: Bool
+          #if LazyInitializableByDefault
+            lazyInitializableByDefault = true
+          #else
+            lazyInitializableByDefault = false
+          #endif
+          let isLazyInitializableColumn =
+            isLazyInitializable
+            ?? (lazyInitializableByDefault && defaultValue == nil && !isColumnGroup)
           if let primaryKey, primaryKey.identifier == identifier {
             var property = property
             for attributeIndex in property.attributes.indices {
@@ -979,6 +1008,43 @@ extension TableMacro: MemberMacro {
             if let type = binding.typeAnnotation?.type.asOptionalType() {
               binding.typeAnnotation?.type = type
             }
+            property.bindings = [binding]
+            draftProperties.append(
+              DeclSyntax(
+                property
+                  .with(\.bindingSpecifier, .keyword(.var, trailingTrivia: .space))
+                  .removingAccessors()
+                  .rewritten(selfRewriter)
+              )
+            )
+          } else if isLazyInitializableColumn,
+            !isColumnGroup,
+            let optionalType = binding.typeAnnotation?.type.asOptionalType()
+          {
+            var property = property
+            for attributeIndex in property.attributes.indices {
+              guard
+                var attribute = property.attributes[attributeIndex].as(AttributeSyntax.self)?
+                  .trimmed,
+                attribute.attributeName.as(IdentifierTypeSyntax.self)?.name.text == "Column",
+                case .argumentList(var arguments) = attribute.arguments
+              else { continue }
+              for argumentIndex in arguments.indices {
+                var argument = arguments[argumentIndex]
+                defer { arguments[argumentIndex] = argument }
+                if argument.label?.text == "as",
+                  var expression = argument.expression.as(MemberAccessExprSyntax.self)
+                {
+                  expression.base = "\(expression.base)?"
+                  argument.expression = ExprSyntax(expression)
+                }
+              }
+              attribute.arguments = .argumentList(arguments)
+              property.attributes[attributeIndex] = .attribute(attribute)
+            }
+            property = property.trimmed
+            var binding = binding
+            binding.typeAnnotation?.type = optionalType
             property.bindings = [binding]
             draftProperties.append(
               DeclSyntax(
@@ -1087,6 +1153,9 @@ extension TableMacro: MemberMacro {
               expansionFailed = true
 
             case .some(let label) where label.text == "generated":
+              expansionFailed = true
+
+            case .some(let label) where label.text == "lazyInitializable":
               expansionFailed = true
 
             case let argument?:
