+++ Tests/StructuredQueriesMacrosTests/TableMacroTests.swift
@@ -185,191 +185,6 @@ extension SnapshotTests {
       }
     }
 
-    @Test func comment() {
-      assertMacro {
-        """
-        @Table
-        struct User {
-          /// The user's identifier.
-          let id: /* TODO: UUID */Int // Primary key
-          /// The user's email.
-          var email: String? = ""  // TODO: Should this be non-optional?
-          /// The user's age.
-          var age: Int
-        }
-        """
-      } expansion: {
-        #"""
-        struct User {
-          @StructuredQueries._ColumnCheck(Int.self)
-          /// The user's identifier.
-          let id: /* TODO: UUID */Int // Primary key
-          @StructuredQueries._ColumnCheck(String?.self)
-          /// The user's email.
-          var email: String? = ""  // TODO: Should this be non-optional?
-          @StructuredQueries._ColumnCheck(Int.self)
-          /// The user's age.
-          var age: Int
-
-          public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition, StructuredQueriesCore.PrimaryKeyedTableDefinition {
-            public typealias QueryValue = User
-            public typealias PrimaryKey = Int
-            public let id = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            @StructuredQueries._PrimaryKeyDefault public var primaryKey = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            public let email = StructuredQueriesCore._TableColumn<QueryValue, String?>.for("email", keyPath: \QueryValue.email, default: "")
-            public let age = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("age", keyPath: \QueryValue.age)
-            public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-              var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-              allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.email._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.age._allColumns)
-              return allColumns
-            }
-            public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-              var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-              writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.email._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.age._writableColumns)
-              return writableColumns
-            }
-            public var queryFragment: QueryFragment {
-              "\(self.id), \(self.email), \(self.age)"
-            }
-          }
-
-          public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-            public typealias QueryValue = User
-            public let allColumns: [any StructuredQueriesCore.QueryExpression]
-            public init(
-              id: some StructuredQueriesCore.QueryExpression<Int>,
-              email: some StructuredQueriesCore.QueryExpression<String?> = String?(queryOutput: ""),
-              age: some StructuredQueriesCore.QueryExpression<Int>
-            ) {
-              var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-              allColumns.append(contentsOf: id._allColumns)
-              allColumns.append(contentsOf: email._allColumns)
-              allColumns.append(contentsOf: age._allColumns)
-              self.allColumns = allColumns
-            }
-          }
-          struct Draft: StructuredQueriesCore.TableDraft, StructuredQueriesCore.PartialSelectStatement {
-            public typealias PrimaryTable = User
-            @StructuredQueries._ColumnCheck(Int?.self)
-            var id: /* TODO: UUID */ Int? // Primary key
-            @StructuredQueries._ColumnCheck(String?.self)
-            var email: String? = ""
-            @StructuredQueries._ColumnCheck(Int.self)
-            var age: Int
-
-            public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition {
-              public typealias QueryValue = Draft
-              public let id = StructuredQueriesCore._TableColumn<QueryValue, Int?>.for("id", keyPath: \QueryValue.id, default: nil)
-              public let email = StructuredQueriesCore._TableColumn<QueryValue, String?>.for("email", keyPath: \QueryValue.email, default: "")
-              public let age = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("age", keyPath: \QueryValue.age)
-              public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-                var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-                allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.email._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.age._allColumns)
-                return allColumns
-              }
-              public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-                var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-                writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.email._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.age._writableColumns)
-                return writableColumns
-              }
-              public var queryFragment: QueryFragment {
-                "\(self.id), \(self.email), \(self.age)"
-              }
-            }
-
-            public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-              public typealias QueryValue = Draft
-              public let allColumns: [any StructuredQueriesCore.QueryExpression]
-              public init(
-                id: some StructuredQueriesCore.QueryExpression<Int?> = Int?(queryOutput: nil),
-                email: some StructuredQueriesCore.QueryExpression<String?> = String?(queryOutput: ""),
-                age: some StructuredQueriesCore.QueryExpression<Int>
-              ) {
-                var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-                allColumns.append(contentsOf: id._allColumns)
-                allColumns.append(contentsOf: email._allColumns)
-                allColumns.append(contentsOf: age._allColumns)
-                self.allColumns = allColumns
-              }
-            }
-
-            public typealias QueryValue = Self
-
-            public typealias From = Swift.Never
-
-            public nonisolated static var columns: TableColumns {
-              TableColumns()
-            }
-
-            public nonisolated static var _columnWidth: Swift.Int {
-              var columnWidth = 0
-              columnWidth += Int?._columnWidth
-              columnWidth += String?._columnWidth
-              columnWidth += Int._columnWidth
-              return columnWidth
-            }
-          }
-        }
-
-        nonisolated extension Draft {
-          nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            self.id = try decoder.decode() ?? nil
-            self.email = try decoder.decode() ?? ""
-            let age = try decoder.decode(\QueryValue.age)
-            guard let age else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.age = age
-          }
-          nonisolated init(_ other: PrimaryTable) {
-            self.id = other.id
-            self.email = other.email
-            self.age = other.age
-          }
-        }
-
-        nonisolated extension User: StructuredQueriesCore.Table, StructuredQueriesCore.PrimaryKeyedTable, StructuredQueriesCore.PartialSelectStatement {
-          public typealias QueryValue = Self
-          public typealias From = Swift.Never
-          public nonisolated static var columns: TableColumns {
-            TableColumns()
-          }
-          public nonisolated static var _columnWidth: Int {
-            var columnWidth = 0
-            columnWidth += Int._columnWidth
-            columnWidth += String?._columnWidth
-            columnWidth += Int._columnWidth
-            return columnWidth
-          }
-          public nonisolated static var tableName: String {
-            "users"
-          }
-          public nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            let id = try decoder.decode(\QueryValue.id)
-            self.email = try decoder.decode() ?? ""  // TODO: Should this be non-optional?
-            let age = try decoder.decode(\QueryValue.age)
-            guard let id else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let age else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.id = id
-            self.age = age
-          }
-        }
-        """#
-      }
-    }
-
     @Test func tableName() {
       assertMacro {
         """
@@ -1061,177 +876,6 @@ extension SnapshotTests {
       }
     }
 
-    @Test func columnGeneratedPrimaryKeyedTable() throws {
-      assertMacro {
-        """
-        @Table struct User {
-          let id: Int
-          var name: String
-          @Column(generated: .stored)
-          let generated: Int
-        }
-        """
-      } expansion: {
-        #"""
-        struct User {
-          @StructuredQueries._ColumnCheck(Int.self)
-          let id: Int
-          @StructuredQueries._ColumnCheck(String.self)
-          var name: String
-          let generated: Int
-
-          public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition, StructuredQueriesCore.PrimaryKeyedTableDefinition {
-            public typealias QueryValue = User
-            public typealias PrimaryKey = Int
-            public let id = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            @StructuredQueries._PrimaryKeyDefault public var primaryKey = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            public let name = StructuredQueriesCore._TableColumn<QueryValue, String>.for("name", keyPath: \QueryValue.name)
-            public let generated = StructuredQueriesCore.GeneratedColumn<QueryValue, Int>("generated", keyPath: \QueryValue.generated)
-            public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-              var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-              allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.name._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.generated._allColumns)
-              return allColumns
-            }
-            public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-              var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-              writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.name._writableColumns)
-              return writableColumns
-            }
-            public var queryFragment: QueryFragment {
-              "\(self.id), \(self.name), \(self.generated)"
-            }
-          }
-
-          public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-            public typealias QueryValue = User
-            public let allColumns: [any StructuredQueriesCore.QueryExpression]
-            public init(
-              id: some StructuredQueriesCore.QueryExpression<Int>,
-              name: some StructuredQueriesCore.QueryExpression<String>,
-              generated: some StructuredQueriesCore.QueryExpression<Int>
-            ) {
-              var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-              allColumns.append(contentsOf: id._allColumns)
-              allColumns.append(contentsOf: name._allColumns)
-              allColumns.append(contentsOf: generated._allColumns)
-              self.allColumns = allColumns
-            }
-          }
-          struct Draft: StructuredQueriesCore.TableDraft, StructuredQueriesCore.PartialSelectStatement {
-            public typealias PrimaryTable = User
-            @StructuredQueries._ColumnCheck(Int?.self)
-            var id: Int?
-            @StructuredQueries._ColumnCheck(String.self)
-            var name: String
-
-            public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition {
-              public typealias QueryValue = Draft
-              public let id = StructuredQueriesCore._TableColumn<QueryValue, Int?>.for("id", keyPath: \QueryValue.id, default: nil)
-              public let name = StructuredQueriesCore._TableColumn<QueryValue, String>.for("name", keyPath: \QueryValue.name)
-              public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-                var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-                allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.name._allColumns)
-                return allColumns
-              }
-              public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-                var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-                writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.name._writableColumns)
-                return writableColumns
-              }
-              public var queryFragment: QueryFragment {
-                "\(self.id), \(self.name)"
-              }
-            }
-
-            public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-              public typealias QueryValue = Draft
-              public let allColumns: [any StructuredQueriesCore.QueryExpression]
-              public init(
-                id: some StructuredQueriesCore.QueryExpression<Int?> = Int?(queryOutput: nil),
-                name: some StructuredQueriesCore.QueryExpression<String>
-              ) {
-                var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-                allColumns.append(contentsOf: id._allColumns)
-                allColumns.append(contentsOf: name._allColumns)
-                self.allColumns = allColumns
-              }
-            }
-
-            public typealias QueryValue = Self
-
-            public typealias From = Swift.Never
-
-            public nonisolated static var columns: TableColumns {
-              TableColumns()
-            }
-
-            public nonisolated static var _columnWidth: Swift.Int {
-              var columnWidth = 0
-              columnWidth += Int?._columnWidth
-              columnWidth += String._columnWidth
-              return columnWidth
-            }
-          }
-        }
-
-        nonisolated extension Draft {
-          nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            self.id = try decoder.decode() ?? nil
-            let name = try decoder.decode(\QueryValue.name)
-            guard let name else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.name = name
-          }
-          nonisolated init(_ other: PrimaryTable) {
-            self.id = other.id
-            self.name = other.name
-          }
-        }
-
-        nonisolated extension User: StructuredQueriesCore.Table, StructuredQueriesCore.PrimaryKeyedTable, StructuredQueriesCore.PartialSelectStatement {
-          public typealias QueryValue = Self
-          public typealias From = Swift.Never
-          public nonisolated static var columns: TableColumns {
-            TableColumns()
-          }
-          public nonisolated static var _columnWidth: Int {
-            var columnWidth = 0
-            columnWidth += Int._columnWidth
-            columnWidth += String._columnWidth
-            columnWidth += Int._columnWidth
-            return columnWidth
-          }
-          public nonisolated static var tableName: String {
-            "users"
-          }
-          public nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            let id = try decoder.decode(\QueryValue.id)
-            let name = try decoder.decode(\QueryValue.name)
-            let generated = try decoder.decode(\QueryValue.generated)
-            guard let id else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let name else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let generated else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.id = id
-            self.name = name
-            self.generated = generated
-          }
-        }
-        """#
-      }
-    }
-
     @Test func computed() {
       assertMacro {
         """
@@ -1802,167 +1446,6 @@ extension SnapshotTests {
             guard let name else {
               throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
             }
-            self.name = name
-          }
-        }
-        """#
-      }
-    }
-
-    @Test func ephemeralFieldPrimaryKeyedTable() {
-      assertMacro {
-        """
-        @Table struct SyncUp {
-          let id: Int
-          var name: String
-          @Ephemeral
-          var computed: Int
-        }
-        """
-      } expansion: {
-        #"""
-        struct SyncUp {
-          @StructuredQueries._ColumnCheck(Int.self)
-          let id: Int
-          @StructuredQueries._ColumnCheck(String.self)
-          var name: String
-          var computed: Int
-
-          public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition, StructuredQueriesCore.PrimaryKeyedTableDefinition {
-            public typealias QueryValue = SyncUp
-            public typealias PrimaryKey = Int
-            public let id = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            @StructuredQueries._PrimaryKeyDefault public var primaryKey = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            public let name = StructuredQueriesCore._TableColumn<QueryValue, String>.for("name", keyPath: \QueryValue.name)
-            public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-              var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-              allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.name._allColumns)
-              return allColumns
-            }
-            public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-              var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-              writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.name._writableColumns)
-              return writableColumns
-            }
-            public var queryFragment: QueryFragment {
-              "\(self.id), \(self.name)"
-            }
-          }
-
-          public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-            public typealias QueryValue = SyncUp
-            public let allColumns: [any StructuredQueriesCore.QueryExpression]
-            public init(
-              id: some StructuredQueriesCore.QueryExpression<Int>,
-              name: some StructuredQueriesCore.QueryExpression<String>
-            ) {
-              var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-              allColumns.append(contentsOf: id._allColumns)
-              allColumns.append(contentsOf: name._allColumns)
-              self.allColumns = allColumns
-            }
-          }
-          struct Draft: StructuredQueriesCore.TableDraft, StructuredQueriesCore.PartialSelectStatement {
-            public typealias PrimaryTable = SyncUp
-            @StructuredQueries._ColumnCheck(Int?.self)
-            var id: Int?
-            @StructuredQueries._ColumnCheck(String.self)
-            var name: String
-
-            public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition {
-              public typealias QueryValue = Draft
-              public let id = StructuredQueriesCore._TableColumn<QueryValue, Int?>.for("id", keyPath: \QueryValue.id, default: nil)
-              public let name = StructuredQueriesCore._TableColumn<QueryValue, String>.for("name", keyPath: \QueryValue.name)
-              public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-                var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-                allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.name._allColumns)
-                return allColumns
-              }
-              public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-                var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-                writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.name._writableColumns)
-                return writableColumns
-              }
-              public var queryFragment: QueryFragment {
-                "\(self.id), \(self.name)"
-              }
-            }
-
-            public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-              public typealias QueryValue = Draft
-              public let allColumns: [any StructuredQueriesCore.QueryExpression]
-              public init(
-                id: some StructuredQueriesCore.QueryExpression<Int?> = Int?(queryOutput: nil),
-                name: some StructuredQueriesCore.QueryExpression<String>
-              ) {
-                var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-                allColumns.append(contentsOf: id._allColumns)
-                allColumns.append(contentsOf: name._allColumns)
-                self.allColumns = allColumns
-              }
-            }
-
-            public typealias QueryValue = Self
-
-            public typealias From = Swift.Never
-
-            public nonisolated static var columns: TableColumns {
-              TableColumns()
-            }
-
-            public nonisolated static var _columnWidth: Swift.Int {
-              var columnWidth = 0
-              columnWidth += Int?._columnWidth
-              columnWidth += String._columnWidth
-              return columnWidth
-            }
-          }
-        }
-
-        nonisolated extension Draft {
-          nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            self.id = try decoder.decode() ?? nil
-            let name = try decoder.decode(\QueryValue.name)
-            guard let name else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.name = name
-          }
-          nonisolated init(_ other: PrimaryTable) {
-            self.id = other.id
-            self.name = other.name
-          }
-        }
-
-        nonisolated extension SyncUp: StructuredQueriesCore.Table, StructuredQueriesCore.PrimaryKeyedTable, StructuredQueriesCore.PartialSelectStatement {
-          public typealias QueryValue = Self
-          public typealias From = Swift.Never
-          public nonisolated static var columns: TableColumns {
-            TableColumns()
-          }
-          public nonisolated static var _columnWidth: Int {
-            var columnWidth = 0
-            columnWidth += Int._columnWidth
-            columnWidth += String._columnWidth
-            return columnWidth
-          }
-          public nonisolated static var tableName: String {
-            "syncUps"
-          }
-          public nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            let id = try decoder.decode(\QueryValue.id)
-            let name = try decoder.decode(\QueryValue.name)
-            guard let id else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let name else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.id = id
             self.name = name
           }
         }
@@ -3454,173 +2937,6 @@ extension SnapshotTests {
       }
     }
 
-    @Test func willSet() {
-      assertMacro {
-        """
-        @Table
-        struct Foo {
-          var id: Int {
-            willSet { print(newValue) }
-          }
-          var name: String {
-            willSet { print(newValue) }
-          }
-        }
-        """
-      } expansion: {
-        #"""
-        struct Foo {
-          @StructuredQueries._ColumnCheck(Int.self)
-          var id: Int {
-            willSet { print(newValue) }
-          }
-          @StructuredQueries._ColumnCheck(String.self)
-          var name: String {
-            willSet { print(newValue) }
-          }
-
-          public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition, StructuredQueriesCore.PrimaryKeyedTableDefinition {
-            public typealias QueryValue = Foo
-            public typealias PrimaryKey = Int
-            public let id = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            @StructuredQueries._PrimaryKeyDefault public var primaryKey = StructuredQueriesCore._TableColumn<QueryValue, Int>.for("id", keyPath: \QueryValue.id)
-            public let name = StructuredQueriesCore._TableColumn<QueryValue, String>.for("name", keyPath: \QueryValue.name)
-            public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-              var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-              allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.name._allColumns)
-              return allColumns
-            }
-            public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-              var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-              writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.name._writableColumns)
-              return writableColumns
-            }
-            public var queryFragment: QueryFragment {
-              "\(self.id), \(self.name)"
-            }
-          }
-
-          public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-            public typealias QueryValue = Foo
-            public let allColumns: [any StructuredQueriesCore.QueryExpression]
-            public init(
-              id: some StructuredQueriesCore.QueryExpression<Int>,
-              name: some StructuredQueriesCore.QueryExpression<String>
-            ) {
-              var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-              allColumns.append(contentsOf: id._allColumns)
-              allColumns.append(contentsOf: name._allColumns)
-              self.allColumns = allColumns
-            }
-          }
-          struct Draft: StructuredQueriesCore.TableDraft, StructuredQueriesCore.PartialSelectStatement {
-            public typealias PrimaryTable = Foo
-            @StructuredQueries._ColumnCheck(Int?.self)
-            var id: Int?
-            @StructuredQueries._ColumnCheck(String.self)
-            var name: String
-
-            public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition {
-              public typealias QueryValue = Draft
-              public let id = StructuredQueriesCore._TableColumn<QueryValue, Int?>.for("id", keyPath: \QueryValue.id, default: nil)
-              public let name = StructuredQueriesCore._TableColumn<QueryValue, String>.for("name", keyPath: \QueryValue.name)
-              public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-                var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-                allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.name._allColumns)
-                return allColumns
-              }
-              public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-                var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-                writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.name._writableColumns)
-                return writableColumns
-              }
-              public var queryFragment: QueryFragment {
-                "\(self.id), \(self.name)"
-              }
-            }
-
-            public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-              public typealias QueryValue = Draft
-              public let allColumns: [any StructuredQueriesCore.QueryExpression]
-              public init(
-                id: some StructuredQueriesCore.QueryExpression<Int?> = Int?(queryOutput: nil),
-                name: some StructuredQueriesCore.QueryExpression<String>
-              ) {
-                var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-                allColumns.append(contentsOf: id._allColumns)
-                allColumns.append(contentsOf: name._allColumns)
-                self.allColumns = allColumns
-              }
-            }
-
-            public typealias QueryValue = Self
-
-            public typealias From = Swift.Never
-
-            public nonisolated static var columns: TableColumns {
-              TableColumns()
-            }
-
-            public nonisolated static var _columnWidth: Swift.Int {
-              var columnWidth = 0
-              columnWidth += Int?._columnWidth
-              columnWidth += String._columnWidth
-              return columnWidth
-            }
-          }
-        }
-
-        nonisolated extension Draft {
-          nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            self.id = try decoder.decode() ?? nil
-            let name = try decoder.decode(\QueryValue.name)
-            guard let name else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.name = name
-          }
-          nonisolated init(_ other: PrimaryTable) {
-            self.id = other.id
-            self.name = other.name
-          }
-        }
-
-        nonisolated extension Foo: StructuredQueriesCore.Table, StructuredQueriesCore.PrimaryKeyedTable, StructuredQueriesCore.PartialSelectStatement {
-          public typealias QueryValue = Self
-          public typealias From = Swift.Never
-          public nonisolated static var columns: TableColumns {
-            TableColumns()
-          }
-          public nonisolated static var _columnWidth: Int {
-            var columnWidth = 0
-            columnWidth += Int._columnWidth
-            columnWidth += String._columnWidth
-            return columnWidth
-          }
-          public nonisolated static var tableName: String {
-            "foos"
-          }
-          public nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            let id = try decoder.decode(\QueryValue.id)
-            let name = try decoder.decode(\QueryValue.name)
-            guard let id else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let name else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.id = id
-            self.name = name
-          }
-        }
-        """#
-      }
-    }
-
     @Test func advanced() {
       assertMacro {
         """
@@ -4447,349 +3763,5 @@ extension SnapshotTests {
       }
     }
 
-    @Test func customPrimaryKey() {
-      assertMacro {
-        """
-        @Table
-        private struct ReminderWithList {
-          @Column(primaryKey: true)
-          let reminderID: Reminder.ID
-          let reminderTitle: String
-          let remindersListTitle: String
-        }
-        """
-      } expansion: {
-        #"""
-        private struct ReminderWithList {
-          let reminderID: Reminder.ID
-          @StructuredQueries._ColumnCheck(String.self)
-          let reminderTitle: String
-          @StructuredQueries._ColumnCheck(String.self)
-          let remindersListTitle: String
-
-          public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition, StructuredQueriesCore.PrimaryKeyedTableDefinition {
-            public typealias QueryValue = ReminderWithList
-            public typealias PrimaryKey = Reminder.ID
-            public let reminderID = StructuredQueriesCore.TableColumn<QueryValue, Reminder.ID>("reminderID", keyPath: \QueryValue.reminderID)
-            @StructuredQueries._PrimaryKeyDefault public var primaryKey = StructuredQueriesCore.TableColumn<QueryValue, Reminder.ID>("reminderID", keyPath: \QueryValue.reminderID)
-            public let reminderTitle = StructuredQueriesCore._TableColumn<QueryValue, String>.for("reminderTitle", keyPath: \QueryValue.reminderTitle)
-            public let remindersListTitle = StructuredQueriesCore._TableColumn<QueryValue, String>.for("remindersListTitle", keyPath: \QueryValue.remindersListTitle)
-            public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-              var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-              allColumns.append(contentsOf: QueryValue.columns.reminderID._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.reminderTitle._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.remindersListTitle._allColumns)
-              return allColumns
-            }
-            public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-              var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-              writableColumns.append(contentsOf: QueryValue.columns.reminderID._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.reminderTitle._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.remindersListTitle._writableColumns)
-              return writableColumns
-            }
-            public var queryFragment: QueryFragment {
-              "\(self.reminderID), \(self.reminderTitle), \(self.remindersListTitle)"
-            }
-          }
-
-          public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-            public typealias QueryValue = ReminderWithList
-            public let allColumns: [any StructuredQueriesCore.QueryExpression]
-            public init(
-              reminderID: some StructuredQueriesCore.QueryExpression<Reminder.ID>,
-              reminderTitle: some StructuredQueriesCore.QueryExpression<String>,
-              remindersListTitle: some StructuredQueriesCore.QueryExpression<String>
-            ) {
-              var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-              allColumns.append(contentsOf: reminderID._allColumns)
-              allColumns.append(contentsOf: reminderTitle._allColumns)
-              allColumns.append(contentsOf: remindersListTitle._allColumns)
-              self.allColumns = allColumns
-            }
-          }
-          fileprivate struct Draft: StructuredQueriesCore.TableDraft, StructuredQueriesCore.PartialSelectStatement {
-            public typealias PrimaryTable = ReminderWithList
-            var reminderID: Reminder.ID?
-            @StructuredQueries._ColumnCheck(String.self)
-            let reminderTitle: String
-            @StructuredQueries._ColumnCheck(String.self)
-            let remindersListTitle: String
-
-            public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition {
-              public typealias QueryValue = Draft
-              public let reminderID = StructuredQueriesCore.TableColumn<QueryValue, Reminder.ID?>("reminderID", keyPath: \QueryValue.reminderID, default: nil)
-              public let reminderTitle = StructuredQueriesCore._TableColumn<QueryValue, String>.for("reminderTitle", keyPath: \QueryValue.reminderTitle)
-              public let remindersListTitle = StructuredQueriesCore._TableColumn<QueryValue, String>.for("remindersListTitle", keyPath: \QueryValue.remindersListTitle)
-              public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-                var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-                allColumns.append(contentsOf: QueryValue.columns.reminderID._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.reminderTitle._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.remindersListTitle._allColumns)
-                return allColumns
-              }
-              public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-                var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-                writableColumns.append(contentsOf: QueryValue.columns.reminderID._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.reminderTitle._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.remindersListTitle._writableColumns)
-                return writableColumns
-              }
-              public var queryFragment: QueryFragment {
-                "\(self.reminderID), \(self.reminderTitle), \(self.remindersListTitle)"
-              }
-            }
-
-            public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-              public typealias QueryValue = Draft
-              public let allColumns: [any StructuredQueriesCore.QueryExpression]
-              public init(
-                reminderID: some StructuredQueriesCore.QueryExpression<Reminder.ID?> = Reminder.ID?(queryOutput: nil),
-                reminderTitle: some StructuredQueriesCore.QueryExpression<String>,
-                remindersListTitle: some StructuredQueriesCore.QueryExpression<String>
-              ) {
-                var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-                allColumns.append(contentsOf: reminderID._allColumns)
-                allColumns.append(contentsOf: reminderTitle._allColumns)
-                allColumns.append(contentsOf: remindersListTitle._allColumns)
-                self.allColumns = allColumns
-              }
-            }
-
-            public typealias QueryValue = Self
-
-            public typealias From = Swift.Never
-
-            public nonisolated static var columns: TableColumns {
-              TableColumns()
-            }
-
-            public nonisolated static var _columnWidth: Swift.Int {
-              var columnWidth = 0
-              columnWidth += Reminder.ID?._columnWidth
-              columnWidth += String._columnWidth
-              columnWidth += String._columnWidth
-              return columnWidth
-            }
-          }
-        }
-
-        nonisolated extension Draft {
-          fileprivate nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            self.reminderID = try decoder.decode() ?? nil
-            let reminderTitle = try decoder.decode(\QueryValue.reminderTitle)
-            let remindersListTitle = try decoder.decode(\QueryValue.remindersListTitle)
-            guard let reminderTitle else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let remindersListTitle else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.reminderTitle = reminderTitle
-            self.remindersListTitle = remindersListTitle
-          }
-          fileprivate nonisolated init(_ other: PrimaryTable) {
-            self.reminderID = other.reminderID
-            self.reminderTitle = other.reminderTitle
-            self.remindersListTitle = other.remindersListTitle
-          }
-        }
-
-        nonisolated extension ReminderWithList: StructuredQueriesCore.Table, StructuredQueriesCore.PrimaryKeyedTable, StructuredQueriesCore.PartialSelectStatement {
-          public typealias QueryValue = Self
-          public typealias From = Swift.Never
-          public nonisolated static var columns: TableColumns {
-            TableColumns()
-          }
-          public nonisolated static var _columnWidth: Int {
-            var columnWidth = 0
-            columnWidth += Reminder.ID._columnWidth
-            columnWidth += String._columnWidth
-            columnWidth += String._columnWidth
-            return columnWidth
-          }
-          public nonisolated static var tableName: String {
-            "reminderWithLists"
-          }
-          public nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            let reminderID = try decoder.decode(\QueryValue.reminderID)
-            let reminderTitle = try decoder.decode(\QueryValue.reminderTitle)
-            let remindersListTitle = try decoder.decode(\QueryValue.remindersListTitle)
-            guard let reminderID else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let reminderTitle else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let remindersListTitle else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.reminderID = reminderID
-            self.reminderTitle = reminderTitle
-            self.remindersListTitle = remindersListTitle
-          }
-        }
-        """#
-      }
-    }
-
-    @Test func composite() {
-      assertMacro {
-        """
-        @Table
-        private struct Metadata: Identifiable {
-          let id: MetadataID
-          var userModificationDate: Date
-        }
-        """
-      } expansion: {
-        #"""
-        private struct Metadata: Identifiable {
-          @StructuredQueries._ColumnCheck(MetadataID.self)
-          let id: MetadataID
-          @StructuredQueries._ColumnCheck(Date.self)
-          var userModificationDate: Date
-
-          public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition, StructuredQueriesCore.PrimaryKeyedTableDefinition {
-            public typealias QueryValue = Metadata
-            public typealias PrimaryKey = MetadataID
-            public let id = StructuredQueriesCore._TableColumn<QueryValue, MetadataID>.for("id", keyPath: \QueryValue.id)
-            @StructuredQueries._PrimaryKeyDefault public var primaryKey = StructuredQueriesCore._TableColumn<QueryValue, MetadataID>.for("id", keyPath: \QueryValue.id)
-            public let userModificationDate = StructuredQueriesCore._TableColumn<QueryValue, Date>.for("userModificationDate", keyPath: \QueryValue.userModificationDate)
-            public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-              var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-              allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-              allColumns.append(contentsOf: QueryValue.columns.userModificationDate._allColumns)
-              return allColumns
-            }
-            public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-              var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-              writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-              writableColumns.append(contentsOf: QueryValue.columns.userModificationDate._writableColumns)
-              return writableColumns
-            }
-            public var queryFragment: QueryFragment {
-              "\(self.id), \(self.userModificationDate)"
-            }
-          }
-
-          public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-            public typealias QueryValue = Metadata
-            public let allColumns: [any StructuredQueriesCore.QueryExpression]
-            public init(
-              id: some StructuredQueriesCore.QueryExpression<MetadataID>,
-              userModificationDate: some StructuredQueriesCore.QueryExpression<Date>
-            ) {
-              var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-              allColumns.append(contentsOf: id._allColumns)
-              allColumns.append(contentsOf: userModificationDate._allColumns)
-              self.allColumns = allColumns
-            }
-          }
-          fileprivate struct Draft: StructuredQueriesCore.TableDraft, StructuredQueriesCore.PartialSelectStatement {
-            public typealias PrimaryTable = Metadata
-            @StructuredQueries._ColumnCheck(MetadataID?.self)
-            var id: MetadataID?
-            @StructuredQueries._ColumnCheck(Date.self)
-            var userModificationDate: Date
-
-            public nonisolated struct TableColumns: StructuredQueriesCore.TableDefinition {
-              public typealias QueryValue = Draft
-              public let id = StructuredQueriesCore._TableColumn<QueryValue, MetadataID?>.for("id", keyPath: \QueryValue.id, default: nil)
-              public let userModificationDate = StructuredQueriesCore._TableColumn<QueryValue, Date>.for("userModificationDate", keyPath: \QueryValue.userModificationDate)
-              public static var allColumns: [any StructuredQueriesCore.TableColumnExpression] {
-                var allColumns: [any StructuredQueriesCore.TableColumnExpression] = []
-                allColumns.append(contentsOf: QueryValue.columns.id._allColumns)
-                allColumns.append(contentsOf: QueryValue.columns.userModificationDate._allColumns)
-                return allColumns
-              }
-              public static var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] {
-                var writableColumns: [any StructuredQueriesCore.WritableTableColumnExpression] = []
-                writableColumns.append(contentsOf: QueryValue.columns.id._writableColumns)
-                writableColumns.append(contentsOf: QueryValue.columns.userModificationDate._writableColumns)
-                return writableColumns
-              }
-              public var queryFragment: QueryFragment {
-                "\(self.id), \(self.userModificationDate)"
-              }
-            }
-
-            public nonisolated struct Selection: StructuredQueriesCore.TableExpression {
-              public typealias QueryValue = Draft
-              public let allColumns: [any StructuredQueriesCore.QueryExpression]
-              public init(
-                id: some StructuredQueriesCore.QueryExpression<MetadataID?> = MetadataID?(queryOutput: nil),
-                userModificationDate: some StructuredQueriesCore.QueryExpression<Date>
-              ) {
-                var allColumns: [any StructuredQueriesCore.QueryExpression] = []
-                allColumns.append(contentsOf: id._allColumns)
-                allColumns.append(contentsOf: userModificationDate._allColumns)
-                self.allColumns = allColumns
-              }
-            }
-
-            public typealias QueryValue = Self
-
-            public typealias From = Swift.Never
-
-            public nonisolated static var columns: TableColumns {
-              TableColumns()
-            }
-
-            public nonisolated static var _columnWidth: Swift.Int {
-              var columnWidth = 0
-              columnWidth += MetadataID?._columnWidth
-              columnWidth += Date._columnWidth
-              return columnWidth
-            }
-          }
-        }
-
-        nonisolated extension Draft {
-          fileprivate nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            self.id = try decoder.decode() ?? nil
-            let userModificationDate = try decoder.decode(\QueryValue.userModificationDate)
-            guard let userModificationDate else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.userModificationDate = userModificationDate
-          }
-          fileprivate nonisolated init(_ other: PrimaryTable) {
-            self.id = other.id
-            self.userModificationDate = other.userModificationDate
-          }
-        }
-
-        nonisolated extension Metadata: StructuredQueriesCore.Table, StructuredQueriesCore.PrimaryKeyedTable, StructuredQueriesCore.PartialSelectStatement {
-          public typealias QueryValue = Self
-          public typealias From = Swift.Never
-          public nonisolated static var columns: TableColumns {
-            TableColumns()
-          }
-          public nonisolated static var _columnWidth: Int {
-            var columnWidth = 0
-            columnWidth += MetadataID._columnWidth
-            columnWidth += Date._columnWidth
-            return columnWidth
-          }
-          public nonisolated static var tableName: String {
-            "metadatas"
-          }
-          public nonisolated init(decoder: inout some StructuredQueriesCore.QueryDecoder) throws {
-            let id = try decoder.decode(\QueryValue.id)
-            let userModificationDate = try decoder.decode(\QueryValue.userModificationDate)
-            guard let id else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            guard let userModificationDate else {
-              throw StructuredQueriesCore.QueryDecodingError.missingRequiredColumn
-            }
-            self.id = id
-            self.userModificationDate = userModificationDate
-          }
-        }
-        """#
-      }
-    }
   }
 }
